requirements.cpu.txt
bs4==0.0.2
hf-xet==1.1.10
nltk==3.9.1
numpy==2.2.6
pandas==2.3.2
qdrant-client==1.15.1
requests==2.32.5
scikit-learn==1.7.2
scipy==1.15.3
aiohttp==3.12.15
fastapi==0.118.0
google-genai==1.40.0
uvicorn==0.37.0
redis==7.0.0
python-dotenv==1.1.1
sentencepiece==0.2.1
pypdf2==3.0.1
python-docx==1.2.0

# Add CPU-only PyTorch explicitly before sentence-transformers
torch==2.5.1+cpu --index-url https://download.pytorch.org/whl/cpu
sentence-transformers==5.1.1